+++ labcodes_answer/lab8/user/libs/syscall.h
@@ -16,7 +16,7 @@ int sys_gettime(void);
 struct stat;
 struct dirent;
 
-int sys_open(const char *path, uint32_t open_flags);
+int sys_open(const char *path, uint64_t open_flags);
 int sys_close(int fd);
 int sys_read(int fd, void *base, size_t len);
 int sys_write(int fd, void *base, size_t len);
@@ -26,7 +26,7 @@ int sys_fsync(int fd);
 int sys_getcwd(char *buffer, size_t len);
 int sys_getdirentry(int fd, struct dirent *dirent);
 int sys_dup(int fd1, int fd2);
-void sys_lab6_set_priority(uint32_t priority); //only for lab6
+void sys_lab6_set_priority(uint64_t priority); //only for lab6
 
 
 #endif /* !__USER_LIBS_SYSCALL_H__ */
